+++ src/engines/niseci/full/mod.rs
@@ -103,16 +103,16 @@ pub fn calculate_niseci(
     let (x3, criteri_x3) = x3.expect("calc_niseci() returned earlier on Err match");
 
     if let Some(ref crit) = criteri_x3 {
-        let submetriche_map_x3 = crit.get_submetriche_map();
+        let submetriche_map_x3 = crit.get_ref_submetriche_map();
         for (key, val) in submetriche_map_x3 {
-            let classi_eta = val.get_classi_eta();
+            let classi_eta = val.get_ref_classi_eta();
             let specie = key.clone();
             let densita_stimata = -1.0; //TODO: check if this is correct
             let subvalue_a = val.get_criterio_a();
             let subvalue_b = val.get_criterio_b();
             let val = ValoriIntermediSpecieNISECI {
                 #[expect(deprecated)]
-                classi_eta,
+                classi_eta: classi_eta.clone(),
                 #[expect(deprecated)]
                 densita_stimata,
                 #[expect(deprecated)]
@@ -206,17 +206,17 @@ fn get_valori_intermedi_specie(
 ) -> HashMap<String, ValoriIntermediSpecieNISECI> {
     let mut valori_intermedi_specie: HashMap<String, ValoriIntermediSpecieNISECI> = HashMap::new();
 
-    let submetriche_map = criteri.get_submetriche_map();
+    let submetriche_map = criteri.get_ref_submetriche_map();
     for (key, val) in submetriche_map.iter() {
         let criteri_x2_a = val.get_metriche_x2_a();
-        let classi_eta = val.get_classi_eta();
+        let classi_eta = val.get_ref_classi_eta();
         let specie = key.clone();
-        let densita_stimata = val.get_metriche_x2_b().get_densita_stimata();
-        let quantita_stimata = val.get_metriche_x2_b().get_quantita_stimata();
-        let x2_b = val.get_metriche_x2_b().get_x2_b();
+        let densita_stimata = val.get_ref_metriche_x2_b().get_densita_stimata();
+        let quantita_stimata = val.get_ref_metriche_x2_b().get_quantita_stimata();
+        let x2_b = val.get_ref_metriche_x2_b().get_x2_b();
         let val = ValoriIntermediSpecieNISECI {
             #[expect(deprecated)]
-            classi_eta,
+            classi_eta: classi_eta.clone(),
             #[expect(deprecated)]
             densita_stimata,
             #[expect(deprecated)]